task 0

+++ 0x1E-search_algorithms/0-linear.c
@@ -2,18 +2,18 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t index = 0;
+	int index = 0;
 
 	if (value)
-	for (; index < size; index ++)
+	for (; index < (int)size; index ++)
 	{
 		if (array[index] == value)
 		{
-			return ((int)index);
+			return (index);
 		}
 		else 
 		{
-			printf("Value checked array[%d] = [%d]\n", (int)index, array[index]);
+			printf("Value checked array[%d] = [%d]\n", index, array[index]);
 		}
 	}
